acho q foi

+++ KING_Event_Team/KING_Event_Team/src/main/java/br/com/fatec/controller/GerenIngressosController.java
@@ -10,6 +10,7 @@ import br.com.fatec.model.Evento;
 import br.com.fatec.model.Ingresso;
 import java.net.URL;
 import java.sql.SQLException;
+import java.util.Optional;
 import java.util.ResourceBundle;
 import javafx.beans.property.ReadOnlyStringWrapper;
 import javafx.collections.FXCollections;
@@ -19,6 +20,7 @@ import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
@@ -115,17 +117,24 @@ public class GerenIngressosController implements Initializable
             App.mensagem("AVISO", "Selecione um Ingresso!", Alert.AlertType.WARNING);
             return;
         }
+        Alert alerta = new Alert(Alert.AlertType.CONFIRMATION);
+        alerta.setTitle("CONFIRMAÇÃO");
+        alerta.setHeaderText("Deseja excluir este item?");
+        Optional<ButtonType> resultado = alerta.showAndWait();
+        if(resultado.isPresent() && resultado.get() == ButtonType.OK){
         
-        ingresso.setEvento(evento);
-        try
-        {
-            dao.remover(ingresso);
-            App.mensagem("SUCESSO", "Ingresso removido com sucesso!");
-            preencherTabela();
-        }
-        catch(SQLException ex)
-        {
-            App.mensagem("ERRO", "Erro ao remover ingresso: " + ex.getMessage(), Alert.AlertType.ERROR);
+        
+            ingresso.setEvento(evento);
+            try
+            {
+                dao.remover(ingresso);
+                App.mensagem("SUCESSO", "Ingresso removido com sucesso!");
+                preencherTabela();
+            }
+            catch(SQLException ex)
+            {
+                App.mensagem("ERRO", "Erro ao remover ingresso: " + ex.getMessage(), Alert.AlertType.ERROR);
+            }
         }
     }
